Moved models.py to upc_models.py and updated imports
